🔧 chore(ansible-lint.yml): remove yaml[octal-values] from skip_list ✨ feat(ansible roles): wrap mode values in quotes to improve readability and consistency The yaml[octal-values] rule was removed from the skip_list in the ansible-lint.yml file. The mode values in the ansible roles were wrapped in quotes to improve readability and consistency.
🐛 fix(roles): wrap mode values in quotes to avoid Ansible warnings
The mode values in the Ansible tasks were wrapped in quotes to avoid warnings from Ansible. This is a minor fix that improves the readability of the code and ensures that the tasks run without warnings.

🐛 fix(main.yml): wrap mode values in quotes to avoid Ansible deprecation warning
The mode values in the Ansible tasks were wrapped in quotes to avoid deprecation warnings. This is a minor fix that ensures the playbook runs without warnings.