Fix linter for Consul (#370)
* 🔥 chore(ansible-lint.yml): remove ../roles/consul from exclude_paths
The ../roles/consul path was previously excluded from ansible-lint checks, but it is no longer necessary as the issue it was addressing has been resolved.

* 🎨 style(consul): change variable name from `consul_autopilot_cleanup_dead_Servers` to `consul_autopilot_cleanup_dead_servers` for consistency
The variable name `consul_autopilot_cleanup_dead_Servers` has been changed to `consul_autopilot_cleanup_dead_servers` to improve consistency with the naming conventions used in the rest of the codebase.

* 🐛 fix(dnsmasq.yml): remove unnecessary jinja2 brackets in when statement
The jinja2 brackets in the when statement are unnecessary and can be removed. This commit removes them to improve readability and consistency with other tasks in the playbook.

* 🐛 fix(encrypt_gossip.yml): add pipefail option to shell command to prevent silent errors
The `set -o pipefail` option was added to the shell command to prevent silent errors that may occur when the `grep` command fails to find the "encrypt" key in the `config.json` file. This ensures that any errors that occur during the execution of the command are caught and reported.