Cleanup of ansible-lint.yml file (#376)
* 🔥 chore(ansible-lint.yml): remove trailing-spaces rule from skip_list
The trailing-spaces rule was removed from the skip_list in order to enforce the rule and ensure that there are no trailing spaces in the YAML files.

* 🔥 chore(ansible-lint.yml): remove schema[tasks] from skip_list
The schema[tasks] rule was removed from the skip_list as it is no longer necessary.

* 🔥 chore(ansible-lint.yml): remove skip_list items 'role-name[path]' and 'run-once[task]'
The 'role-name[path]' and 'run-once[task]' items were removed from the skip_list as they are no longer needed. The 'role-name[path]' item was removed because it is redundant with the 'role-name' item. The 'run-once[task]' item was removed because it is not used in the project.

* 🔥 chore(ansible-lint.yml): remove commented out lines
The commented out lines were not providing any value and were just cluttering the file.