Enable no-free-form rule and refactor tasks (#377)
* 🔥 chore(ansible-lint.yml): remove no-free-form rule from skip_list
The no-free-form rule was removed from the skip_list in order to enable it. This rule checks that all tasks have a `task` key and that it is not a free-form task.

* 🎨 style(ansible): reformat service task in ansible-role-firewall and set_fact tasks in consul role
The service task in ansible-role-firewall was reformatted to use a dictionary for the service name and state. This improves readability and consistency with other tasks in the playbook. The set_fact tasks in the consul role were also reformatted to use a dictionary for the variable assignment. This improves readability and consistency with other tasks in the playbook.