Removal of fqcn[action] rule from ansible-lint skip_list (#378)
* 🔥 chore(ansible-lint.yml): remove fqcn[action] rule from skip_list
The fqcn[action] rule was removed from the skip_list as it is no longer necessary.

* 🚚 chore(configs): update Ansible modules to use community.general and community.docker collections
The Ansible modules used in the playbook have been updated to use the community.general and community.docker collections. This is because the modules have been moved to these collections and are no longer available in the core Ansible distribution. This change ensures that the playbook remains compatible with the latest version of Ansible.

🐛 fix(roles): replace deprecated Ansible modules with their updated counterparts
✨ feat(roles): improve maintainability by using updated Ansible modules
The deprecated Ansible modules have been replaced with their updated counterparts to ensure that the roles continue to work as expected. This improves the maintainability of the roles as the updated modules are more reliable and have better support.

🚚 chore(roles): update Ansible module names to use fully qualified names
The Ansible modules used in the roles have been updated to use fully qualified names. This is to ensure that the correct modules are used and to avoid any potential conflicts with other modules. The updated modules are as follows:
- community.postgresql.postgresql_schema
- community.postgresql.postgresql_user
- ansible.posix.authorized_key
- ansible.posix.mount
- ansible.posix.sysctl
- community.general.timezone
- community.general.make

* 🔨 chore(README.md): update minimum supported Ansible version to 2.11.0
🔨 chore(config_pgcluster.yml): update minimal Ansible version to 2.11.0
🔨 chore(deploy_pgcluster.yml): update minimal Ansible version to 2.11.0
🔨 chore(meta/main.yml): update minimum supported Ansible version to 2.11.0
The minimum supported Ansible version has been updated to 2.11.0 in all relevant files. This is done to ensure compatibility with the latest version of Ansible and to take advantage of its new features and improvements.

* 🆕 chore(runtime.yml): add runtime.yml file with ansible version requirement
This commit adds a new file named runtime.yml which specifies the minimum version of Ansible required to run the application. This file is important as it helps ensure that the application is run on a compatible version of Ansible.

* 📝 chore(runtime.yml): add newline at end of file
This commit adds a newline at the end of the runtime.yml file to improve readability and ensure consistency with other files in the repository.