Parallelize PostgreSQL upgrade on replicas using RSync
This commit introduces a performance improvement to the task "Upgrade the PostgreSQL on the Replica (using rsync --hard-links)" by parallelizing the rsync operation using 'xargs'.

By executing rsync in parallel on multiple secondary servers, we can significantly reduce the time required to upgrade PostgreSQL on the replicas.

The updated code should ensure that the PostgreSQL update task is performed using rsync in asynchronous mode and wait for it to complete to make sure that the update was completed successfully.