Make inventory group 'balancers' optional in Ansible tasks
With this commit, we've adjusted our Ansible tasks to treat the 'balancers' group as optional.
Previously, if this group wasn't defined in the inventory file, it would result in a task failure. We've addressed this issue by utilizing the default filter, which handles the 'balancers' group as an empty list if not defined, thus preventing any associated errors.

Please note that this change doesn't impact the 'postgres_cluster' group, which remains as a required group in the inventory.

Issue: https://github.com/vitabaks/postgresql_cluster/issues/388