Automate 'archive_command' configuration for pgBackRest and WAL-G (#394)
* Automate 'archive_command' configuration for pgBackRest

This commit includes an Ansible task that automates the configuration of the 'archive_command' for pgBackRest. The task ensures that 'archive_command' is set to the value of 'pgbackrest_archive_command'. This operation is performed only when 'pgbackrest_install' is defined and set to true, and 'archive_command' is undefined or its value is not equal to 'pgbackrest_archive_command' variable.

* Automate 'archive_command' configuration for WAL-G

This commit includes an Ansible task that automates the configuration of the 'archive_command' for WAL-G. The task ensures that 'archive_command' is set to the value of 'wal_g_archive_command'. This operation is performed only when 'wal_g_install' is defined and set to true, and 'archive_command' is undefined or its value is not equal to 'wal_g_archive_command' variable.