README: Update the ansible installation method
Unfortunately, most Linux distributions' package managers default to installing older versions of Ansible. To circumvent this, we recommend the use of pip for the installation of Ansible. The pip method installs more recent versions, providing us with the features necessary for a successful deployment. 

This is due to our reliance on features and modules that are not supported in older versions.