pgBackRest: Ensure 'repo1-host' and 'repo1-host-user' are set correctly
This commit adds a check for the existence of the 'repo1-host' and 'repo1-host-user' options when 'pgbackrest_repo_host' is defined. If these options are not set, the configuration is adjusted automatically. This is crucial for the proper functioning of pgBackRest when using a dedicated backup server.

Issue: https://github.com/vitabaks/postgresql_cluster/issues/393