Remove compatibility with Ansible versions older than 2.11.0 (#403)
This commit is aimed at improving the maintainability and usability of our Ansible scripts. Specifically, this involves:

1. Removing support for older Ansible versions: We are discontinuing support for Ansible versions prior to 2.11.0. By doing so, we eliminate the need to write and maintain extra code for backward compatibility, which often complicates the codebase and can introduce potential bugs.
2. Embracing newer functionalities: Dropping support for older versions allows us to fully utilize the advanced features and improvements offered in recent Ansible releases, thereby making our scripts more efficient and powerful.

While this change may require users to upgrade their Ansible installation, it's a step towards keeping our tools up-to-date and as effective as possible.

To mitigate potential disruption, we recommend all users to upgrade their Ansible installations to version 2.11.0 or newer.