Added pgbouncer authentication via auth_user and auth_query (#401)
This commit brings a significant improvement to the database's authentication process through pgbouncer, transitioning from the auth_file method to auth_user and auth_query.

Previously, we relied on the auth_file method, which required maintaining and updating a userlist.txt file each time a user was added to or removed from the database. This method was less flexible and efficient, as it required manual management.

The new implementation utilizes auth_user and auth_query, automating and simplifying the process. The auth_user in pgbouncer allows a single specified user (pgbouncer_auth_username) to query the user's password for authentication using auth_query. This removes the necessity to manually manage a userlist.txt file, thus streamlining the database users and pgbouncer pool management.