📝 docs: Add new test file pg_probackup.yml for testing pg_probackup command 🔧 refactor(vars/main.yml): Replace hardcoded pg_probackup command with a variable
The new test file pg_probackup.yml is added to test the pg_probackup command. This is to ensure that the command is correctly formed and works as expected.

In vars/main.yml, the hardcoded pg_probackup command is replaced with a variable. This change makes the command more flexible and easier to manage. Now, the command can be easily updated or modified by just changing the variable value.