🔧 refactor(haproxy.tmpl.yml): update task names and variables for better clarity 🔀 feat(haproxy.tmpl.yml): add test case for when cluster_vip is not defined The task names and variables in the haproxy.tmpl.yml file have been updated to better reflect their purpose and improve readability. The inventory_hostname and cluster_vip have been changed to 'una.name' and 'fake.vip.url.com' respectively for testing purposes. A new test case has been added to handle the scenario when the cluster_vip is not defined. This will help ensure the application behaves as expected in different environments and configurations.
🔧 refactor(confd/tasks/main.yml): improve loop readability by renaming loop variable and reformatting loop structure
The loop variable has been renamed from 'item' to 'line_item_2' to provide more context about the data it represents. The loop structure has been reformatted to separate the 'regexp' and 'line' properties into separate lines, improving readability and making it easier to understand the loop's purpose. This change is part of an ongoing effort to improve code quality and maintainability.