🔧 refactor(molecule tests): optimize test execution by adding run_once to tasks 🔍 test(molecule tests): adjust file paths in include_vars tasks for correct file location 🔧 refactor(.gitignore): broaden the ignore rule for tmp directories in molecule tests The changes were made to optimize the execution of molecule tests by adding the run_once directive to tasks that only need to be run once, rather than on every host in the inventory. This will speed up the test execution. The file paths in include_vars tasks were adjusted to correctly locate the required files. The .gitignore file was updated to ignore all tmp directories in the molecule directory, not just those in the tests subdirectory, to prevent unnecessary files from being tracked by git.
🔧 refactor(wal_g_cron_jobs.yml, main.yml): remove unnecessary shared variables inclusion and add run_once to tasks
The shared variables inclusion was removed from wal_g_cron_jobs.yml as it was not necessary for the tasks in the file. The run_once attribute was added to all tasks to ensure that they are only executed once, improving the efficiency of the playbook. The main.yml file was also simplified by removing unnecessary tasks and adding the run_once attribute to the remaining task.