🔧 refactor(custom_wal_dir.yml, main.yml): add run_once to tasks to optimize execution 🔨 fix(main.yml): remove unnecessary task details to simplify the code
The 'run_once' directive has been added to all tasks in the 'custom_wal_dir.yml' file to ensure that these tasks are executed only once, regardless of the number of hosts. This change optimizes the execution of the playbook by avoiding unnecessary repetitions of the same tasks.

In the 'main.yml' file, unnecessary task details such as 'hosts' and 'gather_facts' have been removed to simplify the code. The 'run_once' directive has been added to the task to ensure it is executed only once. This change makes the code cleaner and easier to read.