Add auto_compaction for maintenance etcd
This PR is designed to automate the maintenance of etcd cluster. Namely, the addition of automatic compression of the etcd database.

If you are using this change with an existing etcd cluster, after the configuration change and the automatic compression period has passed, it is recommended to perform a one-time defragmentation. 

It is described in more detail here: https://etcd.io/docs/v3.5/op-guide/maintenance/