Fix (Upgrade): Update pg-path in local pgbackrest.conf
The conditions of the task block were incorrect, which is why the local configuration file was not updated. As a result, we encountered errors archiving WAL and backing up to a dedicated pgBackRest server because the new version of Postgres was indicated there while the old version was still in the local configuration.

This fix ensures that the `pg-path` option will be updated on both the dedicated server and the local database servers.