Download pgvectorscale archive in check_mode
Added the check_mode: false parameter to the task responsible for downloading the pgvectorscale archive using the ansible.builtin.get_url module. This ensures that the task executes even in --check mode, preventing it from being skipped and ensuring the file is downloaded. This change addresses an error where subsequent tasks fail due to the missing downloaded archive.